Update config files to selectively run benchmarks
Now that we have some benchmark groups that don't work well with
older GHC versions (e.g. Dph and Repa), we want to be explicit about
which benchmarks we run.

Also, we now look for GHC head in a standard location, but still
allow it to be overwritten by an environment variable.